# docs/swagger.yaml
basePath: /
definitions:
...
    handler.ErrorDetail:
        properties:
            code:
                type: string
            message:
                type: string
        type: object
    handler.ErrorResponse:
        properties:
            error:
                $ref: '#/definitions/handler.ErrorDetail'
        type: object
...
    handler.MessageResponse:
        properties:
            message:
                type: string
        type: object
...
    handler.RideListResponse:
        properties:
            page:
                type: integer
            per_page:
                type: integer
            rides: {}
            total:
                type: integer
            total_pages:
                type: integer
        type: object
    handler.RideReceiptResponse:
        properties:
            receipt:
                properties:
                    base_fare:
                        type: number
                    distance_fare:
                        type: number
                    surge_multiplier:
                        type: number
                    time_fare:
                        type: number
                    total:
                        type: number
                type: object
        type: object
    handler.RideResponse:
        properties:
            ride:
                $ref: '#/definitions/model.Ride'
        type: object
...
    handler.StubResponse:
        properties:
            message:
                type: string
            status:
                type: string
        type: object
...
    handler.rateRideRequest:
        properties:
            comment:
                type: string
            score:
                type: integer
        required:
            - score
        type: object
...
    handler.rideRequest:
        properties:
            dropoff_address:
                type: string
            dropoff_lat:
                type: number
            dropoff_lng:
                type: number
            pickup_address:
                type: string
            pickup_lat:
                type: number
            pickup_lng:
                type: number
            vehicle_type:
                type: string
        required:
            - dropoff_lat
            - dropoff_lng
            - pickup_lat
            - pickup_lng
        type: object
...
    model.Ride:
        properties:
            accepted_at:
                type: string
            base_fare:
                type: number
            cancellation_fee:
                type: number
            cancelled_at:
                type: string
            completed_at:
                type: string
            created_at:
                type: string
            distance_fare:
                type: number
            driver_arrived_at:
                type: string
            driver_id:
                type: string
            dropoff_address:
                type: string
            dropoff_lat:
                type: number
            dropoff_lng:
                type: number
            id:
                type: string
            pickup_address:
                type: string
            pickup_lat:
                type: number
            pickup_lng:
                type: number
            requested_at:
                type: string
            rider_id:
                type: string
            started_at:
                type: string
            status:
                type: string
            surge_multiplier:
                type: number
            time_fare:
                type: number
            total_fare:
                type: number
            updated_at:
                type: string
            vehicle_type:
                type: string
        type: object
...
host: localhost:8080
info:
    contact: {}
    description: 'Go + Gin ride-hailing backend: auth, rides, driver dispatch, geolocation, navigation, and places.'
    title: Ride-Hailing API
    version: 0.1.0
paths:
...
    /api/v1/rides:
        post:
            consumes:
                - application/json
            description: Creates a ride request and dispatches it to nearby drivers.
            operationId: postRides
            parameters:
                - description: Ride request
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.rideRequest'
            produces:
                - application/json
            responses:
                "201":
                    description: Created
                    schema:
                        $ref: '#/definitions/handler.RideResponse'
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "500":
                    description: Failed to create ride
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Request a new ride
            tags:
                - rides
    /api/v1/rides/{id}:
        get:
            operationId: getRides-ByID
            parameters:
                - description: Ride ID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.RideResponse'
                "404":
                    description: Ride not found
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Get a ride by ID
            tags:
                - rides
    /api/v1/rides/{id}/cancel:
        post:
            operationId: postRides-ByID-cancel
            parameters:
                - description: Ride ID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.RideResponse'
                "400":
                    description: Ride cannot be cancelled
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Cancel a ride
            tags:
                - rides
...
    /api/v1/rides/{id}/rate:
        post:
            consumes:
                - application/json
            description: Submits a 1-5 rating for the other party on a completed ride.
            operationId: postRides-ByID-rate
            parameters:
                - description: Ride ID
                  in: path
                  name: id
                  required: true
                  type: string
                - description: Rating payload
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/handler.rateRideRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.MessageResponse'
                "400":
                    description: Invalid rating
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "404":
                    description: Ride not found
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
                "422":
                    description: Validation error
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Rate a completed ride
            tags:
                - rides
    /api/v1/rides/{id}/receipt:
        get:
            operationId: getRides-ByID-receipt
            parameters:
                - description: Ride ID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.RideReceiptResponse'
                "404":
                    description: Ride not found
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: Get the fare breakdown for a ride
            tags:
                - rides
    /api/v1/rides/{id}/tip:
        post:
            operationId: postRides-ByID-tip
            parameters:
                - description: Ride ID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.StubResponse'
            security:
                - BearerAuth: []
            summary: Tip a driver (stub)
            tags:
                - rides
    /api/v1/rides/current:
        get:
            description: Returns the active ride for the current rider or driver, or `null`.
            operationId: getRides-current
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.RideResponse'
            security:
                - BearerAuth: []
            summary: Get the current ride for the authenticated user
            tags:
                - rides
    /api/v1/rides/history:
        get:
            description: Returns a paginated list of past rides for the current rider or driver.
            operationId: getRides-history
            parameters:
                - default: 1
                  description: Page number
                  in: query
                  name: page
                  type: integer
                - default: 20
                  description: Items per page (max 50)
                  in: query
                  name: per_page
                  type: integer
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/handler.RideListResponse'
                "500":
                    description: Query failed
                    schema:
                        $ref: '#/definitions/handler.ErrorResponse'
            security:
                - BearerAuth: []
            summary: List ride history
            tags:
                - rides
...
securityDefinitions:
    BearerAuth:
        description: 'JWT access token. Include as `Authorization: Bearer <token>`.'
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"